# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/spells": {
      "get": {
        "operationId": "SpellsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field (prefix with '-' for descending order)",
            "schema": {
              "example": "-createdAt",
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter spells by name",
            "schema": {
              "example": "Fireball",
              "type": "string"
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Filter by language code (ISO 2 code)",
            "schema": {
              "example": "en",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spells found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IPaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Spell"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get a collection of paginated spells",
        "tags": [
          "Spells"
        ]
      },
      "post": {
        "operationId": "SpellsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpellDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spell created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Spell"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Create a new spell",
        "tags": [
          "Spells"
        ]
      }
    },
    "/spells/{id}": {
      "get": {
        "operationId": "SpellsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the spell to get",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "lang",
            "required": true,
            "in": "path",
            "description": "The ISO 2 code of translation",
            "schema": {
              "example": "en",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spell found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Spell"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error while fetching spell #ID: Id is not a valid mongoose id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Spell #ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "410": {
            "description": "Spell #ID has been deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Get a spell by ID",
        "tags": [
          "Spells"
        ]
      },
      "patch": {
        "operationId": "SpellsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the spell to update",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpellDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spell updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Spell"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Spell #ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "410": {
            "description": "Spell #ID has been deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Update a spell by ID",
        "tags": [
          "Spells"
        ]
      },
      "delete": {
        "operationId": "SpellsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the spell to delete",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spell #ID deleted",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Spell"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error while fetching spell #ID: Id is not a valid mongoose id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Spell #ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "410": {
            "description": "Spell #ID has been deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Delete a spell by ID",
        "tags": [
          "Spells"
        ]
      }
    },
    "/spells/{id}/translations": {
      "get": {
        "operationId": "SpellsController_getTranslations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the spell to get translations for",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spell translations found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SpellTranslationsListDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error while fetching spell #ID: Id is not a valid mongoose id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Spell #ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "410": {
            "description": "Spell #ID has been deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Get all available translations for a spell",
        "tags": [
          "Spells"
        ]
      }
    },
...
  },
  "info": {
    "title": "Codex API",
    "description": "Codex API Documentation",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:9000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Spell": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "number",
            "example": 1
          },
          "languages": {
            "example": [
              "en"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "translations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SpellContent"
            }
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "example": null
          }
        },
        "required": [
          "tag",
          "languages",
          "translations",
          "deletedAt"
        ]
      },
      "IResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Spell #507f1f77bcf86cd799439011 found in 5ms"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "offset": {
            "type": "number",
            "example": 10
          },
          "totalItems": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "page",
          "offset",
          "totalItems"
        ]
      },
      "IPaginatedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Spell #507f1f77bcf86cd799439011 found in 5ms"
          },
          "data": {
            "type": "object"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": [
          "message",
          "data",
          "pagination"
        ]
      },
...
      "SpellTranslationsListDto": {
        "type": "object",
        "properties": {
          "spellId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011",
            "description": "Spell ID"
          },
          "tag": {
            "type": "number",
            "example": 1,
            "description": "Spell tag"
          },
          "translations": {
            "description": "List of available translations",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellTranslationSummaryDto"
            }
          }
        },
        "required": [
          "spellId",
          "tag",
          "translations"
        ]
      },
...
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://httpstatuses.io/400"
          },
          "title": {
            "type": "string",
            "example": "Bad request"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "instance": {
            "type": "string",
            "example": "/spells"
          },
          "detail": {
            "type": "string",
            "example": "Validation failed"
          },
          "invalid-params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParamDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "detail"
        ]
      },
      "UpdateSpellDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "number",
            "example": 1,
            "description": "0: Homebrew, 1: Certified by Chariot",
            "enum": [
              0,
              1
            ]
          }
        },
        "required": [
          "tag"
        ]
      },
      "CreateSpellContentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fireball"
          },
          "description": {
            "type": "string",
            "example": "A bright streak flashes from your pointing finger to a point you choose within range."
          },
          "level": {
            "type": "number",
            "example": 3
          },
          "school": {
            "type": "string",
            "example": "Evocation"
          },
          "castingTime": {
            "type": "string",
            "example": "1 action"
          },
          "range": {
            "type": "string",
            "example": "150 feet"
          },
          "components": {
            "example": [
              "V",
              "S",
              "M"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "duration": {
            "type": "string",
            "example": "Instantaneous"
          },
          "effectType": {
            "type": "number",
            "example": 1
          },
          "damage": {
            "type": "string",
            "example": "8d6"
          },
          "healing": {
            "type": "string",
            "example": "Heal 4d8 + spellcasting modifier"
          }
        },
        "required": [
          "name",
          "description",
          "level",
          "school",
          "castingTime",
          "range",
          "components",
          "duration",
          "effectType",
          "damage",
          "healing"
        ]
      },
      "CreateSpellDto": {
        "type": "object",
        "properties": {
          "lang": {
            "type": "string",
            "example": "en"
          },
          "spellContent": {
            "$ref": "#/components/schemas/CreateSpellContentDto"
          }
        },
        "required": [
          "lang",
          "spellContent"
        ]
      },
...
    }
  }
}